Added adding books

+++ app/models/BookEntry.scala
@@ -2,6 +2,6 @@ package models
 
 import java.time.LocalDateTime
 
-case class BookEntry (id: Int, user_id: Int, isbn: String,
+case class BookEntry (id: Long, user_id: Long, isbn: String,
                       created_at: LocalDateTime, status: String,
                       pages_read: Integer)
